fix epoll stop data race (#201)
A stop call (aws_event_loop_stop) on an epoll event loop updates the same stop
task regardless of whether the stop task is currently scheduled. This can lead
to null pointer dereferences.

This bug is not present in the kqueue or IOCP event-loops, which use a
mutex-protected status flag.