Improved Posix Large File Support (LFS)
Mildly breaking API change!!! All uses of `aws_off_t` replaced by `int64_t`

### ISSUE:
Bizarre behavior on 32bit Raspberry Pi OS when aws_input_stream_seek() was called.

### CAUSE:
Conditional logic in header files is bad.

The size of `off_t` on Posix is conditional, based on various defines when headers are included.

It's possible for .c files in different libraries to disagree on the size of `off_t`, which means they disagree on the size of `aws_off_t`, which leads to stack corruption when functions taking `aws_off_t` are called.

### CHANGES:
1) Always use `int64_t` for offsets, instead of `aws_off_t`. Deal with 32 vs 64 bit offsets in .c files we control, not headers included by external code. I will fix up downstream libraries in followup PRs.
2) AwsCFlags now attempts to enable LFS on 32 bit system, see https://github.com/awslabs/aws-c-common/pull/808