Beginnings of hooking up PKCS#11 to s2n_ctx.
Implemented:
- aws_pkcs11_lib_find_slot_with_token()
- aws_pkcs11_lib_open_session()
- aws_pkcs11_lib_close_session()
- aws_pkcs11_lib_login_user()

Decided to have a single PKCS#11 session shared by all connections on the aws_tls_ctx, to avoid hitting limits on the number of sessions. We can revisit this if it becomes a bottleneck.

Realizing we to give user control over C_Login() being called, the same way we gave control over C_Initialize(), but we'll tackle that later.