Find PKCS#11 private key / begin s2n integration (#425)
**Design decision:**
Use a single PKCS#11 session for all TLS connections on this s2n_ctx. We do this because PKCS#11 tokens may only support a [limited number of sessions](http://docs.oasis-open.org/pkcs11/pkcs11-ug/v2.40/cn02/pkcs11-ug-v2.40-cn02.html#_Toc406759996). Only one thread can interact with a session at a time, so this is potentially a big bottleneck.

For now, we'll keep it simple, and just use a lock to ensure the session has only 1 user at a time. If this turns out to be a major bottleneck in reality, there are many possible ways to optimize, including:
- put the single session on its own thread outside the event-loops
- 1 session per connection (depending on max supported sessions)
- 1 session per event-loop (depending on max supported sessions)

**Also lots of little helper functions:**
  - `aws_pkcs11_lib_find_slot_with_token()`
  - `aws_pkcs11_lib_open_session()`
  - `aws_pkcs11_lib_close_session()`
  - `aws_pkcs11_lib_login_user()`
  - `aws_pkcs11_lib_find_private_key()`