Move PKCS#11 stuff outside of ifdef block.
This wasn't meant to be inside an `if !defined(AWS_OS_IOS)` block, that was an accident. The PKCS#11 stuff is declared on every platform, but raises AWS_ERROR_UNIMPLEMENTED at runtime if it's not actually supported on a platform. This is meant to reduce the complexity of the bindings.