Fix aws-crt-swift by moving PKCS#11 headers out of include/ folder.
It seems like Swift is trying to individually process these files because they're under include/, even if we we add these files to the `exclude` list.

The official PKCS#11 headers cannot be parse individually. The user is required to `#define` a few things before `#include`ing them.

Moving these headers out from under the the include/ folder fixes the issue