aws_future macros generate normal, not inline, functions
Rust was having trouble calling these functions. aws_future didn't really NEED the functions to be inline. It wasn't really a performance thing, I just liked how inline let there be a single DECLARATION in a header.

Solution is to have separate DECLARATION macro to go in a header, and an IMPLEMENTATION macro to go in a C file. Now they're normal functions.