aws_future functions no longer inline (#577)
Rust was having trouble calling `aws_future_bool` functions, due to them only being available as inline.

These functions were inline so that the we'd only need 1 declaration macro in a header. It wasn't a performance thing.

Fix is to have the macros generate normal non-inline functions, with a DECLARATION macro that goes in a header, and an IMPLEMENTATION macro that goes in a C file.